# docs/spec/openapi.yaml
...
openapi: 3.1.1
info:
  title: Data Veracity Assurance
  description: |-
    This is the API specification of the
    [data veracity assurance building block (DVA)](https://github.com/Prometheus-X-association/data-veracity).
  contact:
    email: bpeter@edu.bme.hu
  version: 0.5.1
servers:
  - url: http://localhost:9099
    description: Test environment – VLA manager service
  - url: http://localhost:9091
    description: Test environment – provider side
  - url: http://localhost:9092
    description: Test environment – consumer side
  - url: http://localhost:9090
    description: Local development server running on default port
  - url: '{server}'
    description: Custom parameterizable server
    variables:
      server:
        default: http://localhost:9090
        description: Custom server URL
...
paths:
...
  /evaluate:
    post:
      tags: [Evaluation]
      summary: Evaluate a requirement against some data
      description: >
        Takes some requirement (that could be part of a VLA) and a piece of data in
        the request body.
        The response is the result of evaluating the requirement on that data.
        This endpoint is useful for testing requirements in the process of building
        VLAs.
      operationId: evaluate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
      responses:
        '200':
          $ref: '#/components/responses/Evaluated'
        '400':
          $ref: '#/components/responses/InvalidData'
  /evaluate/from-template:
    post:
      tags: [Evaluation]
      summary: Render a template and evaluate the result against some  data
      description: >
        Takes a template ID and some data in the request body.
        The template with the given ID is rendered internally and the resulting
        requirement is evaluated on the data passed.
        This endpoint is useful for testing requirements in the process of building
        VLAs.
      operationId: evaluteFromTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationFromTemplateRequest'
      responses:
        '200':
          $ref: '#/components/responses/Evaluated'
        '400':
          $ref: '#/components/responses/InvalidData'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    EvaluationRequest:
      type: object
      required: [requirement, data]
      additionalProperties: false
      properties:
        requirement:
          $ref: https://raw.githubusercontent.com/bitol-io/open-data-contract-standard/refs/heads/main/schema/odcs-json-schema-latest.json#/$defs/DataQualityCustom
        data:
          type: object
      example:
        requirement:
          engine: JQ
          implementation: >-
            if .actor == 'Expected Actor'
            then { success: true, details: 'Actor name is correct' }
            else { success: false, details: 'Unexpected actor name' }
            end
        data:
          statement_id: 123
          actor: Unexpected Uriel
          action: read
          object: course_materials/lecturenotes/1
    EvaluationFromTemplateRequest:
      type: object
      required: [templateID, templateModel, data]
      additionalProperties: false
      properties:
        templateID:
          type: string
        templateModel:
          type: object
        data:
          type: object
      example:
        templateID: 74d696f2-b217-4d2f-a0c1-5ff65cd2258d
        templateModel:
          target_field: .actor
          required_value: Expected Actor
        data:
          statement_id: 123
          actor: Unexpected Uriel
          action: read
          object: course_materials/lecturenotes/1
    EvaluationResult:
      $ref: './components.yaml#/schemas/EvaluationResult'
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
  responses:
...
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            Template:
              summary: Template not found
              value:
                type: NOT_FOUND
                message: Could not find VLA template with id 'foo-123'
            AoV:
              summary: Attestation (AoV) not found
              value:
                type: NOT_FOUND
                message: Could not find AoV with id 'bar-123'
            VLA:
              summary: VLA not found
              value:
                type: NOT_FOUND
                message: Could not find VLA with id 'bar-123'
    InvalidData:
      description: >-
        The request body was invalid.
        Please do not retry the request without correcting the payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            Template:
              summary: Template data invalid
              value:
                type: INVALID_DATA
                message: Invalid template data supplied
            TemplateRender:
              summary: Template model to render invalid
              value:
                type: INVALID_DATA
                message: Invalid template render request
            AoV:
              summary: Attestation (AoV) generation payload invalid
              value:
                type: INVALID_DATA
                message: Invalid AoV data supplied
...
    Evaluated:
      description: The requirement has been evaluated successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EvaluationResult'